safari-534.53-branch/LayoutTests/editing/inserting/5994480-2.html

<div id="description">This tests to see if deleting an input element that starts at the start of a block adds an extra placeholder.  You should see a single placeholder in the block below.</div>
<div id="edit" contentEditable="true"><input type="text"><br></div>
<script>
if (window.layoutTestController)
    window.layoutTestController.dumpAsText();
edit = document.getElementById("edit");
edit.focus();

document.execCommand("SelectAll");
document.execCommand("Delete");
if (window.layoutTestController)
    document.body.innerText = edit.innerHTML;
</script>
