LayoutTests/platform/ios-simulator/editing/inserting/5994480-2-expected.txt

This tests to see if deleting an input element that starts at the start of a block adds an extra placeholder.
You should see a single placeholder in the block below.
| <input>
|   type="text"
|   this.value=""
|   <shadow:root>
|     <div>
|       contenteditable="plaintext-only"
| <br>
